+++ c_projects/PalindromeString/Palindrome.c
