+++ c_projects/PalindromeString/ReadME.md
